homework 55 complete

+++ src/main/java/com/example/jobsearch/service/VacancyService.java
@@ -22,6 +22,7 @@ public class VacancyService {
     private final VacancyDao vacancyDao;
     private final EmployerProfileService employerService;
     private final UserService userService;
+    private final CategoryService categoryService;
 
     public List<VacancyDto> getAllVacancies() {
         List<Vacancy> list = vacancyDao.getAllVacancies();
@@ -97,8 +98,13 @@ public class VacancyService {
                 v = vacancyDao.findVacancyById(vacancyDto.getId());
             }
             if (v.isEmpty()) {
-                vacancyDao.save(makeVacancyFromDto(vacancyDto));
-                return new ResponseEntity<>("Vacancy was created successfully", HttpStatus.OK);
+                if (categoryService.getCategory(vacancyDto.getCategory()).isPresent()) {
+                    vacancyDao.save(makeVacancyFromDto(vacancyDto));
+                    return new ResponseEntity<>("Vacancy was created successfully", HttpStatus.OK);
+                } else {
+                    log.warn("Tried to use a category that does not exist: {}", vacancyDto.getCategory());
+                    return new ResponseEntity<>("Category does not exist", HttpStatus.BAD_REQUEST);
+                }
             } else {
                 log.info("Tried to create a vacancy that already exists: {}", vacancyDto.getId());
                 return new ResponseEntity<>("Vacancy already exists", HttpStatus.OK);
@@ -110,22 +116,53 @@ public class VacancyService {
 
     }
 
-    public void editVacancy(VacancyDto vacancyDto, Authentication auth) {
+    public ResponseEntity<?> editVacancy(VacancyDto vacancyDto, Authentication auth) {
         var u = auth.getPrincipal();
         User user = userService.getUserFromAuth(u.toString());
         if (employerService.getUserIdByEmployerId(vacancyDto.getEmployerId()).equalsIgnoreCase(user.getId())) {
-            Vacancy vacancy = makeVacancyFromDto(vacancyDto);
-            vacancyDao.editVacancy(vacancy);
+            if (vacancyDto.getId() == null) {
+                return new ResponseEntity<>("Cannot edit a vacancy without vacancy id", HttpStatus.NOT_FOUND);
+            }
+            var v = vacancyDao.findVacancyById(vacancyDto.getId());
+            if (v.isEmpty()) {
+                log.info("Tried to edit a vacancy that that does not exist: {}", vacancyDto.getId());
+                return new ResponseEntity<>("Vacancy does not exist", HttpStatus.OK);
+            } else {
+                if(categoryService.getCategory(vacancyDto.getCategory()).isPresent()) {
+                    Vacancy vacancy = makeVacancyFromDto(vacancyDto);
+                    vacancyDao.editVacancy(vacancy);
+                    return new ResponseEntity<>("Vacancy edited successfully", HttpStatus.OK);
+                }else {
+                    log.warn("Tried to use a category that does not exist: {}", vacancyDto.getCategory());
+                    return new ResponseEntity<>("Category does not exist", HttpStatus.BAD_REQUEST);
+                }
+            }
+        } else {
+            log.warn("Tried to edit a vacancy of another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to edit a vacancy of another user", HttpStatus.BAD_REQUEST);
         }
     }
 
-    public void deleteVacancy(VacancyDto vacancyDto, Authentication auth) {
+    public ResponseEntity<?> deleteVacancy(VacancyDto vacancyDto, Authentication auth) {
         var u = auth.getPrincipal();
         User user = userService.getUserFromAuth(u.toString());
         if (employerService.getUserIdByEmployerId(vacancyDto.getEmployerId()).equalsIgnoreCase(user.getId())) {
-            log.info("Vacancy was deleted: {}", vacancyDto.getId());
-            vacancyDao.delete(vacancyDto.getId());
+            var v = vacancyDao.findVacancyById(vacancyDto.getId());
+            if (v.isEmpty()) {
+                log.info("Tried to delete a vacancy that does not exist: {}", vacancyDto.getId());
+                return new ResponseEntity<>("Tried to delete a vacancy that does not exist", HttpStatus.OK);
+            } else {
+                vacancyDao.delete(vacancyDto.getId());
+                log.info("Vacancy was deleted: {}", vacancyDto.getId());
+                return new ResponseEntity<>("Vacancy was deleted successfully", HttpStatus.OK);
+
+            }
+
+        } else {
+            log.warn("Tried to delete a vacancy of another user: {}", user.getId());
+            return new ResponseEntity<>("Tried to delete a vacancy of another user", HttpStatus.BAD_REQUEST);
         }
+
     }
 
     public ResponseEntity<?> sortedListVacancies(String sortedCriteria) {
@@ -144,5 +181,14 @@ public class VacancyService {
         }
     }
 
-
+    public ResponseEntity<?> getAllVacanciesByEmployerId(Long employerId) {
+        List<Vacancy> list = vacancyDao.getAllVacanciesByEmployerId(employerId);
+        if (list.isEmpty()) {
+            return new ResponseEntity<>("There are no vacancies by this employer", HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>(list.stream()
+                    .map(this::makeDtoFromVacancy)
+                    .toList(), HttpStatus.OK);
+        }
+    }
 }
